requirements.txt
djangorestframework~=3.9.2
django-redis==4.10.0
Django==2.2.4
structlog~=16.1.0
requests~=2.23.0
psycopg2
django-rq
rq-scheduler
iso3166
